+++ apps/lib/router/routes.dart
@@ -1,0 +1,17 @@
+import 'package:apps/pages/hoge_page.dart';
+import 'package:flutter/material.dart';
+import 'package:go_router/go_router.dart';
+
+part 'routes.g.dart';
+
+@TypedGoRoute<HomeRoute>(
+  path: '/',
+)
+class HomeRoute extends GoRouteData {
+  const HomeRoute();
+
+  @override
+  Widget build(BuildContext context, GoRouterState state) {
+    return const MyHomePage(title: 'Flutter Demo Home Page');
+  }
+}
